pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["jeSoosRemirez <dkhomych0091@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
unidecode = "^1.3.8"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
drf-yasg = ">=1.21.7"
gunicorn = "^22.0.0"
django-cors-headers = "^4.3.1"
django-debug-toolbar = "^4.3.0"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
